#8 Movi las rutas de ideas a ideas.js
